+++ erasure_code/erasure_code_test.c
@@ -187,7 +187,7 @@ gf_gen_decode_matrix(unsigned char *encode_matrix, unsigned char *decode_matrix,
                                 continue;
                         }
                 }
-                if (decode_index[k - 1] + incr >= m) {
+                if (decode_index[k - 1] + incr >= (unsigned int) m) {
                         free(b);
                         free(backup);
                         printf("BAD MATRIX\n");
@@ -671,7 +671,7 @@ main(int argc, char *argv[])
 
                 for (i = 0; i < m; i++) {
 
-                        offset = ubuffs[i] - buffs[i];
+                        offset = (unsigned int) (ubuffs[i] - buffs[i]);
 
                         if (memcmp(buffs[i], temp_buffs[0], offset)) {
                                 printf("Fail rand ualign encode pad start\n");
@@ -688,7 +688,7 @@ main(int argc, char *argv[])
 
                 for (i = 0; i < nerrs; i++) {
 
-                        offset = temp_ubuffs[k + i] - temp_buffs[k + i];
+                        offset = (unsigned int) (temp_ubuffs[k + i] - temp_buffs[k + i]);
                         if (memcmp(temp_buffs[k + i], temp_buffs[0], offset)) {
                                 printf("Fail rand ualign decode pad start\n");
                                 re = -1;
